refactor: 데이터 포맷팅 함수로 이름 변경
