+++ src/utils/index.ts
@@ -1,5 +1,5 @@
 export * from './getDate';
 export * from './router';
-export * from './formatStatusToKR';
+export * from './localeFormatter';
 export * from './queryClient';
 export * from './signout';
